perf(plugin): fragments no longer need to be copied over to the .cache directory because gatsby looks through node_modules
